feat: add generator with server
