+++ libs/modules/app-stories/src/useCases/createApp/CreateAppRequest.generated.ts
@@ -1,0 +1,9 @@
+import gql from 'graphql-tag'
+
+export const CreateAppGql = gql`
+  mutation CreateApp($input: CreateAppRequest!) {
+    createApp(request: $input) {
+      title
+    }
+  }
+`
