+++ libs/tools/generators/graphql/src/main.ts
@@ -1,67 +1,33 @@
-import * as path from 'path'
-import { generate } from '@graphql-codegen/cli'
-import * as glob from 'glob'
-
-const outputFile = 'graphql.ts'
-
-const graphqlQueryPaths = glob.sync(
-  `${process.cwd()}/libs/modules/*@(-stories)/src/useCases/**/*.graphql`,
-)
-
-interface GeneratorOptions {
-  sourceGqlPath: string
-}
-
-const makeGenerator = async (option: GeneratorOptions) => {
-  const { sourceGqlPath } = option
-
-  const useCaseDirectory = path.dirname(sourceGqlPath)
-  const filenameWithoutExtension = path.basename(sourceGqlPath, '.graphql')
-
-  console.log(useCaseDirectory, filenameWithoutExtension)
-
-  return generate(
-    {
-      schema: 'http://localhost:3333/graphql',
-      documents: sourceGqlPath,
-      // type error, need to stub
-      // eslint-ignore next-line
-      hooks: {
-        afterStart: [],
-        beforeDone: [],
-        onWatchTriggered: [],
-        onError: [],
-        afterOneFileWrite: [],
-        afterAllFileWrite: ['npx eslint --fix'],
-        beforeOneFileWrite: [],
-        beforeAllFileWrite: [],
-      },
-      generates: {
-        [`${useCaseDirectory}/${filenameWithoutExtension}.generated.ts`]: {
-          plugins: [
-            {
-              'typescript-document-nodes': {
-                nameSuffix: 'Gql',
-              },
-            },
-          ],
-        },
-      },
-    },
-    true,
-  )
-}
+import { spawn } from 'child_process'
+import { getGenerators } from './generator'
 
 export const main = async () => {
-  const generators = graphqlQueryPaths.map((queryPath) => {
-    return makeGenerator({ sourceGqlPath: queryPath })
+  const startServerCmd = `npx env-cmd -f .env cross-env PORT=4001 \
+    node dist/apps/api/codelab/main.js`
+
+  const codelabApiServer = spawn(startServerCmd, {
+    stdio: 'inherit',
+    shell: true,
   })
 
-  // console.log(generatedFiles)
+  codelabApiServer.on('message', (msg) => {
+    console.log(msg)
+  })
 
-  // fs.writeFile(path.join(process.cwd(), outputFile), output, () => {
-  //   console.log('Outputs generated!')
-  // })
+  const generateCmd = 'npx wait-on http://localhost:4001 && exit 0'
+
+  const generator = spawn(generateCmd, {
+    stdio: 'inherit',
+    shell: true,
+  })
+
+  // Server has started
+  generator.on('close', async (msg) => {
+    await Promise.all(getGenerators())
+
+    // Close server
+    codelabApiServer.kill()
+  })
 }
 
 main()
