+++ package.json
@@ -204,6 +204,7 @@
     "@testing-library/react": "11.2.2",
     "@testing-library/user-event": "12.5.0",
     "@types/bcrypt": "3.0.0",
+    "@types/concurrently": "^5.2.1",
     "@types/cors": "2.8.7",
     "@types/cytoscape": "3.14.11",
     "@types/d3": "5.16.2",
